MVVM封装

+++ app/src/main/java/com/yinshan/livedatademo/utils/SpUtils.java
@@ -3,6 +3,8 @@ package com.yinshan.livedatademo.utils;
 import android.content.Context;
 import android.content.SharedPreferences;
 
+import com.yinshan.livedatademo.AppConfig;
+
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Map;
@@ -17,12 +19,12 @@ public class SpUtils {
     /**
      * 保存数据的方法，我们需要拿到保存数据的具体类型，然后根据类型调用不同的保存方法
      *
-     * @param context
      * @param key
      * @param object
      */
-    public static void put(Context context, String key, Object object) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+    public static void put(String key, Object object) {
+        if (null == AppConfig.getInstance().getCurActivity()) return;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = sp.edit();
         if (object instanceof String) {
             editor.putString(key, (String) object);
@@ -45,13 +47,13 @@ public class SpUtils {
     /**
      * 得到保存数据的方法，我们根据默认值得到保存的数据的具体类型，然后调用相对于的方法获取值
      *
-     * @param context
      * @param key
      * @param defaultObject
      * @return
      */
-    public static Object get(Context context, String key, Object defaultObject) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+    public static Object get(String key, Object defaultObject) {
+        if (null == AppConfig.getInstance().getCurActivity()) return null;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         if (defaultObject instanceof String) {
             return sp.getString(key, (String) defaultObject);
         } else if (defaultObject instanceof Integer) {
@@ -75,7 +77,8 @@ public class SpUtils {
      * @param key
      */
     public static void remove(Context context, String key) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+        if (null == AppConfig.getInstance().getCurActivity()) return;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = sp.edit();
         editor.remove(key);
         SharedPreferencesCompat.apply(editor);
@@ -87,7 +90,8 @@ public class SpUtils {
      * @param context
      */
     public static void clear(Context context) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+        if (null == AppConfig.getInstance().getCurActivity()) return;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = sp.edit();
         editor.clear();
         SharedPreferencesCompat.apply(editor);
@@ -96,23 +100,23 @@ public class SpUtils {
     /**
      * 查询某个key是否已经存在
      *
-     * @param context
      * @param key
      * @return
      */
-    public static boolean contains(Context context, String key) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+    public static boolean contains(String key) {
+        if (null == AppConfig.getInstance().getCurActivity()) return false;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         return sp.contains(key);
     }
 
     /**
      * 返回所有的键值对
      *
-     * @param context
      * @return
      */
-    public static Map<String, ?> getAll(Context context) {
-        SharedPreferences sp = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
+    public static Map<String, ?> getAll() {
+        if (null == AppConfig.getInstance().getCurActivity()) return null;
+        SharedPreferences sp = AppConfig.getInstance().getCurActivity().getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE);
         return sp.getAll();
     }
